# kestra/openapi.yml
openapi: 3.0.1
info:
  title: Kestra
  license:
    name: Apache 2.0
    url: https://raw.githubusercontent.com/kestra-io/kestra/master/LICENSE
  version: 1.0.0
...
paths:
...
  /api/v1/bindings:
    get:
      tags:
      - Bindings
      summary: List bindings
      operationId: list_1
      parameters:
      - name: userId
        in: query
        explode: false
        schema:
          type: string
          nullable: true
      - name: roleId
        in: query
        explode: false
        schema:
          type: string
          nullable: true
      - name: namespace
        in: query
        explode: false
        schema:
          type: string
          nullable: true
      responses:
        "200":
          description: list_1 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Binding"
    post:
      tags:
      - Bindings
      summary: Create binding
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BindingController.CreateBindingRequest"
        required: true
      responses:
        "200":
          description: create 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Binding"
  /api/v1/bindings/{id}:
    get:
      tags:
      - Bindings
      summary: Get binding by id
      operationId: get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: get 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Binding"
    delete:
      tags:
      - Bindings
      summary: Delete binding
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: delete 200 response
...
  /api/v1/{tenant}/namespaces/autocomplete:
    post:
      tags:
      - Namespaces
      summary: List namespaces for autocomplete
      description: "Returns a list of namespaces for use in autocomplete fields, optionally\
        \ allowing to filter by query and ids. Used especially for binding creation."
      operationId: autocompleteNamespaces
      parameters:
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiAutocomplete"
        required: true
      responses:
        "200":
          description: autocompleteNamespaces 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /api/v1/{tenant}/namespaces/search:
    get:
      tags:
      - Namespaces
      summary: Search for namespaces
      operationId: searchNamespaces
      parameters:
      - name: q
        in: query
        description: A string filter
        explode: false
        schema:
          type: string
          nullable: true
      - name: page
        in: query
        description: The current page
        required: true
        explode: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        description: The current page size
        required: true
        explode: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
      - name: sort
        in: query
        description: The sort of current page
        explode: false
        schema:
          type: array
          nullable: true
          items:
            type: string
      - name: existing
        in: query
        description: Return only existing namespace
        explode: false
        schema:
          type: boolean
          nullable: true
          default: false
      - name: forceIncludeIds
        in: query
        description: Namespace ids to always include in results (e.g. newly created)
        explode: false
        schema:
          type: array
          nullable: true
          items:
            type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: searchNamespaces 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResults_Namespace_"
  /api/v1/{tenant}/namespaces/{id}:
    get:
      tags:
      - Namespaces
      summary: Get a namespace
      operationId: getNamespace
      parameters:
      - name: id
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: getNamespace 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NamespaceLight"
  /api/v1/{tenant}/namespaces/{namespace}/dependencies:
    get:
      tags:
      - Flows
      summary: Get flow dependencies
      operationId: getFlowDependenciesFromNamespace
      parameters:
      - name: namespace
        in: path
        description: The flow namespace
        required: true
        schema:
          type: string
      - name: destinationOnly
        in: query
        description: "if true, list only destination dependencies, otherwise list\
          \ also source dependencies"
        required: true
        explode: false
        schema:
          type: boolean
          default: false
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: getFlowDependenciesFromNamespace 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlowTopologyGraph"
...
  /api/v1/{tenant}/namespaces/{namespace}/plugin-defaults:
    get:
      tags:
      - Namespaces
      summary: Get namespace plugin defaults
      operationId: getPluginDefaults
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: getPluginDefaults 200 response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PluginDefault"
    put:
      tags:
      - Namespaces
      summary: Set namespace plugin defaults
      operationId: setPluginDefaults
      parameters:
      - name: namespace
        in: path
        description: The namespace id
        required: true
        schema:
          type: string
      - name: tenant
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PluginDefault"
        required: true
      responses:
        "200":
          description: setPluginDefaults 200 response
          content:
            application/json:
              schema:
                type: object
...
components:
  schemas:
...
    ApiAutocomplete:
      type: object
      properties:
        q:
          type: string
          nullable: true
        ids:
          type: array
          nullable: true
          items:
            type: string
        existingOnly:
          type: boolean
...
    Binding:
      required:
      - id
      - roleId
      type: object
      properties:
        id:
          type: string
        roleId:
          type: string
        userId:
          type: string
          nullable: true
        groupId:
          type: string
          nullable: true
        namespace:
          type: string
          nullable: true
    BindingController.CreateBindingRequest:
      type: object
      properties:
        roleId:
          type: string
        userId:
          type: string
        groupId:
          type: string
        namespace:
          type: string
...
    FlowNode:
      required:
      - uid
      type: object
      properties:
        uid:
          type: string
        namespace:
          type: string
        id:
          type: string
    FlowRelation:
      type: string
      enum:
      - FLOW_TASK
      - FLOW_TRIGGER
...
    FlowTopologyGraph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/FlowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/FlowTopologyGraph.Edge"
    FlowTopologyGraph.Edge:
      type: object
      properties:
        source:
          type: string
        target:
          type: string
        relation:
          $ref: "#/components/schemas/FlowRelation"
...
    NamespaceLight:
      required:
      - id
      type: object
      properties:
        id:
          pattern: "^[a-z0-9][a-z0-9._-]*"
          type: string
...
    PagedResults_Namespace_:
      required:
      - results
      - total
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/NamespaceLight"
        total:
          type: integer
          format: int64
...
    PluginDefault:
      required:
      - type
      type: object
      properties:
        type:
          type: string
        forced:
          type: boolean
        values:
          type: object
          additionalProperties: false
...